refactor: extract pull key status and add validation
This also makes sure the pull key is validated when we initially set it
from the query parameters.
